refactor: collapse StoreLifecycle — deep lifecycle seam
- new src/store-lifecycle.ts owns lastPruneMsByStore, lastJanitorMs, live-set (via setStoresGetter), readdir/stat, handleGitPollution as part of onStoreOpen
- exposes onStoreOpen(storePath,stmts,store) + onAppStart() + onSessionStart() + onClose()
- src/hash-store.ts becomes thin: removes warnedGitWarn/handleGitPollution/runCentralJanitorIfDue, delegates to lifecycle, keeps stores/openings Maps and sets getter
- src/index.ts just forwards hooks (onAppStart/onSessionStart), no longer knows throttling
- throttling 24h, live never deleted, WAL checkpoint on close/janitor, undo TTL, pruneMissing batch 64 preserved

Refs: arch review Candidate 2 (Strong) — Store Tenancy/Lifecycle deepening
