+++ src/index.ts
@@ -34,7 +34,7 @@ import {
 	type SectionOverride,
 } from "./guidance.js";
 import { configDir } from "./paths.js";
-import { runCentralJanitorIfDue } from "./hash-store.js";
+import { onAppStart, onSessionStart } from "./store-lifecycle.js";
 
 /** Cordis plugin name used by loader diagnostics. */
 export const name = "dsh-better-edit";
@@ -140,7 +140,7 @@ export function apply(rootCtx: Context): void {
 	// Central lifecycle janitor — throttled 24h, never deletes live hash (mtime hot)
 	// Runs once at app start and on each session-start (central mode only)
 	queueMicrotask(() => {
-		runCentralJanitorIfDue().catch((e) => rootCtx.logger.warn(`dsh-better-edit: central janitor failed: ${e instanceof Error ? e.message : String(e)}`));
+		onAppStart().catch((e) => rootCtx.logger.warn(`dsh-better-edit: central janitor failed: ${e instanceof Error ? e.message : String(e)}`));
 	});
 	// Warm the hasher once; the per-workspace stores are opened lazily on the
 	// first tool call in each workspace (there is no shared store to prune at
@@ -163,7 +163,7 @@ export function apply(rootCtx: Context): void {
 	const registered = new WeakSet<Agent>();
 	rootCtx.on("agent/session-start", ({ agent }) => {
 		// throttled central janitor on session-start (inside same handler to avoid extra listener)
-		runCentralJanitorIfDue().catch((e) => rootCtx.logger.warn(`dsh-better-edit: central janitor failed: ${e instanceof Error ? e.message : String(e)}`));
+		onSessionStart().catch((e) => rootCtx.logger.warn(`dsh-better-edit: central janitor failed: ${e instanceof Error ? e.message : String(e)}`));
 		if (registered.has(agent)) return;
 		registered.add(agent);
 		try {
